+++ modules/try_split/native/MaxDepthCriteria.h
@@ -15,7 +15,7 @@ public:
 
     virtual TrySplitCriteriaI* Clone() const;
 
-    virtual bool TrySplit(int depth, int numberOfDatapoints) const;
+    virtual bool TrySplit(int depth, double numberOfDatapoints) const;
 private:
     const int mMaxDepth;
 };
