+++ modules/try_split/native/TrySplitCombinedCriteria.cpp
@@ -25,7 +25,7 @@ TrySplitCriteriaI* TrySplitCombinedCriteria::Clone() const
     return clone;
 }
 
-bool TrySplitCombinedCriteria::TrySplit(int depth, int numberOfDatapoints) const
+bool TrySplitCombinedCriteria::TrySplit(int depth, double numberOfDatapoints) const
 {
     bool trySplit = true;
     for (std::vector<TrySplitCriteriaI*>::const_iterator it = mCriterias.begin(); it != mCriterias.end() && trySplit; ++it)
