+++ modules/try_split/native/TrySplitCombinedCriteria.h
@@ -16,7 +16,7 @@ public:
 
     virtual TrySplitCriteriaI* Clone() const;
 
-    virtual bool TrySplit(int depth, int numberOfDatapoints) const;
+    virtual bool TrySplit(int depth, double numberOfDatapoints) const;
 private:
     std::vector<TrySplitCriteriaI*> mCriterias;
 };
