+++ modules/try_split/test/test_min_node_size_criteria.cpp
@@ -6,7 +6,7 @@ BOOST_AUTO_TEST_SUITE( MinNodeSizeCriteriaTests )
 
 BOOST_AUTO_TEST_CASE(test_TrySplit)
 {
-    const int minDatapoints = 3;
+    const float minDatapoints = 3;
     MinNodeSizeCriteria minNumberDatapoints(minDatapoints);
     BOOST_CHECK( !minNumberDatapoints.TrySplit(0, minDatapoints-1));
     BOOST_CHECK( minNumberDatapoints.TrySplit(0, minDatapoints));
@@ -15,7 +15,7 @@ BOOST_AUTO_TEST_CASE(test_TrySplit)
 
 BOOST_AUTO_TEST_CASE(test_Clone)
 {
-    const int minDatapoints = 3;
+    const float minDatapoints = 3;
     TrySplitCriteriaI* minNumberDatapoints = new MinNodeSizeCriteria(minDatapoints);
     TrySplitCriteriaI* clone = minNumberDatapoints->Clone();
     delete minNumberDatapoints;
